requirements.txt
# Core
boto3>=1.34.0
botocore>=1.34.0
aws-cdk-lib>=2.130.0
constructs>=10.3.0

# Strands Agents SDK
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Nova Act
nova-act>=0.1.0

# Data Processing
pydantic>=2.5.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Database
psycopg2-binary>=2.9.9

# Utilities
python-dateutil>=2.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tenacity>=8.2.0
structlog>=24.1.0

# Testing
pytest>=7.4.0
moto>=5.0.0
